requirements.txt
Flask==2.3.3
Flask-SocketIO==5.3.5
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.2
python-dotenv==1.0.0
gunicorn==21.2.0
gevent==23.9.1
gevent-websocket==0.10.1
Pygments==2.16.1
pytest==7.4.0
Werkzeug==2.3.7
psycopg2-binary==2.9.7
redis==5.0.1